vet_kg/requirements.txt
neo4j==5.13.0
pandas==2.1.0
networkx==3.1
matplotlib==3.7.2
python-louvain==0.16
spacy==3.7.2
nltk==3.8.1
beautifulsoup4==4.12.2
pyyaml==6.0.1
tqdm==4.66.1 